# Restructure: Allow to track multiple satellites