--- a/src/correlator.jl
+++ b/src/correlator.jl
@@ -1,17 +1,42 @@
-abstract type AbstractCorrelator{T, TA} end
+abstract type AbstractCorrelator{M} end
 """
 $(SIGNATURES)
 
 EarlyPromptLateCorrelator holding a user defined number of correlation values.
 """
-struct EarlyPromptLateCorrelator{T, TA <: AbstractVector{T}} <: AbstractCorrelator{T, TA}
+struct EarlyPromptLateCorrelator{
+    M,
+    T,
+    TA <: AbstractVector{T},
+    S <: AbstractVector{Int}
+} <: AbstractCorrelator{M}
     accumulators::TA
-    EarlyPromptLateCorrelator{T, TA}(accumulators) where {T, TA <: AbstractVector{T}} = length(accumulators) < 3 ?
-        error("Early-Prompt-Late-Correlator needs at least 3 accumulators") : new{T, TA}(accumulators)
-end
-
-function EarlyPromptLateCorrelator(accumulators::AbstractVector)
-    EarlyPromptLateCorrelator{typeof(first(accumulators)), typeof(accumulators)}(accumulators)
+    shifts::S
+    early_shift::Int
+    prompt_shift::Int
+    late_shift::Int
+    function EarlyPromptLateCorrelator(
+        accumulators::AbstractVector{SVector{M, T}},
+        shifts::S,
+        early_shift::Int,
+        prompt_shift::Int,
+        late_shift::Int,
+    ) where {M, T <: Complex, S}
+        length(accumulators) < 3 ?
+            error("Early-Prompt-Late-Correlator needs at least 3 accumulators") :
+            new{M, SVector{M, T}, typeof(accumulators), S}(accumulators, shifts, early_shift, prompt_shift, late_shift)
+    end
+    function EarlyPromptLateCorrelator(
+        accumulators::AbstractVector{T},
+        shifts::S,
+        early_shift::Int,
+        prompt_shift::Int,
+        late_shift::Int,
+    ) where {T <: Complex, S}
+        length(accumulators) < 3 ?
+            error("Early-Prompt-Late-Correlator needs at least 3 accumulators") :
+            new{1, T, typeof(accumulators), S}(accumulators, shifts, early_shift, prompt_shift, late_shift)
+    end
 end
 
 type_for_num_ants(num_ants::NumAnts{1}) = ComplexF64
@@ -23,26 +48,50 @@ $(SIGNATURES)
 EarlyPromptLateCorrelator constructor without parameters and some default parameters.
 """
 function EarlyPromptLateCorrelator(
+    system::AbstractGNSS,
+    sampling_frequency;
+    preferred_code_shift = 0.5,
+    preferred_early_late_code_shift = 0.5,
     num_ants::NumAnts = NumAnts(1),
-    num_accumulators::NumAccumulators{M} = NumAccumulators(3),
-) where M
+    num_accumulators = NumAccumulators(3),
+)
+    shifts = get_correlator_sample_shifts(
+        system,
+        get_num_accumulators(num_accumulators),
+        sampling_frequency,
+        preferred_code_shift
+    )
+    prompt_shift = length(shifts) >> 1 + 1
+
+    sample_shift = preferred_code_shift_to_sample_shift(
+        preferred_early_late_code_shift,
+        sampling_frequency,
+        system
+    )
+    _, late_shift = findmin(abs, shifts .+ sample_shift)
+    _, early_shift = findmin(abs, shifts .- sample_shift)
+
     EarlyPromptLateCorrelator(
-        zero(SVector{M, type_for_num_ants(num_ants)})
+        get_initial_accumulator(num_ants, num_accumulators),
+        shifts,
+        early_shift,
+        prompt_shift,
+        late_shift
     )
 end
 
-"""
-$(SIGNATURES)
+function get_initial_accumulator(
+    num_ants::NumAnts,
+    num_accumulators::NumAccumulators{M}
+) where M
+    zero(SVector{M, type_for_num_ants(num_ants)})
+end
 
-EarlyPromptLateCorrelator constructor for large number of accumulators.
-"""
-function EarlyPromptLateCorrelator(
+function get_initial_accumulator(
     num_ants::NumAnts,
     num_accumulators::Integer
 )
-    EarlyPromptLateCorrelator(
-        [zero(type_for_num_ants(num_ants)) for i = 1:num_accumulators]
-    )
+    [zero(type_for_num_ants(num_ants)) for i = 1:num_accumulators]
 end
 
 
@@ -51,8 +100,7 @@ $(SIGNATURES)
 
 Get number of antennas from correlator
 """
-get_num_ants(correlator::AbstractCorrelator{Complex{T}}) where {T} = 1
-get_num_ants(correlator::AbstractCorrelator{SVector{N,T}}) where {N,T} = N
+get_num_ants(correlator::AbstractCorrelator{M}) where {M} = M
 
 """
 $(SIGNATURES)
@@ -60,29 +108,29 @@ $(SIGNATURES)
 Get number of accumulators
 """
 get_num_accumulators(correlator::AbstractCorrelator) = size(correlator.accumulators, 1)
+get_num_accumulators(num_accumulators::Int) = num_accumulators
+get_num_accumulators(num_accumulators::NumAccumulators{M}) where M = M
 
 """
 $(SIGNATURES)
 
 Get early correlator index
 """
-@inline get_early_index(correlator_sample_shifts, early_late_index_shift) =
-    get_prompt_index(correlator_sample_shifts) + early_late_index_shift
+@inline get_early_index(corr::EarlyPromptLateCorrelator) = corr.early_shift
 
 """
 $(SIGNATURES)
 
 Get prompt correlator index
 """
-@inline get_prompt_index(correlator_sample_shifts) = findfirst(iszero, correlator_sample_shifts)
+@inline get_prompt_index(corr::EarlyPromptLateCorrelator) = corr.prompt_shift
 
 """
 $(SIGNATURES)
 
 Get late correlator index
 """
-@inline get_late_index(correlator_sample_shifts, early_late_index_shift) =
-    get_prompt_index(correlator_sample_shifts) - early_late_index_shift
+@inline get_late_index(corr::EarlyPromptLateCorrelator) = corr.late_shift
 
 """
 $(SIGNATURES)
@@ -97,8 +145,8 @@ $(SIGNATURES)
 Get a specific accumulator with `index` counted negative for late and
 positive for early accumulators.
 """
-function get_accumulator(correlator::AbstractCorrelator, correlator_sample_shifts, index::Integer)
-    correlator.accumulators[index + get_prompt_index(correlator_sample_shifts)]
+function get_accumulator(correlator::AbstractCorrelator, index::Integer)
+    correlator.accumulators[index + get_prompt_index(correlator)]
 end
 
 """
@@ -106,14 +154,8 @@ $(SIGNATURES)
 
 Get early correlator
 """
-function get_early(
-    correlator::AbstractCorrelator,
-    correlator_sample_shifts,
-    early_late_index_shift
-)
-    correlator.accumulators[
-        get_early_index(correlator_sample_shifts, early_late_index_shift)
-    ]
+function get_early(correlator::EarlyPromptLateCorrelator)
+    correlator.accumulators[get_early_index(correlator)]
 end
 
 """
@@ -121,8 +163,8 @@ $(SIGNATURES)
 
 Get prompt correlator
 """
-function get_prompt(correlator::AbstractCorrelator, correlator_sample_shifts)
-    correlator.accumulators[get_prompt_index(correlator_sample_shifts)]
+function get_prompt(correlator::AbstractCorrelator)
+    correlator.accumulators[get_prompt_index(correlator)]
 end
 
 """
@@ -130,14 +172,23 @@ $(SIGNATURES)
 
 Get late correlator
 """
-function get_late(
-    correlator::AbstractCorrelator,
-    correlator_sample_shifts,
-    early_late_index_shift
-)
-    correlator.accumulators[
-        get_late_index(correlator_sample_shifts, early_late_index_shift)
-    ]
+function get_late(correlator::EarlyPromptLateCorrelator)
+    correlator.accumulators[get_late_index(correlator)]
+end
+
+"""
+$(SIGNATURES)
+
+Update the Correlator
+"""
+function update_accumulator(correlator::EarlyPromptLateCorrelator, accumulators) 
+    EarlyPromptLateCorrelator(
+        accumulators,
+        correlator.shifts,
+        correlator.early_shift,
+        correlator.prompt_shift,
+        correlator.late_shift
+    )
 end
 
 """
@@ -145,8 +196,8 @@ $(SIGNATURES)
 
 Zero the Correlator
 """
-function zero(correlator::T) where T <: AbstractCorrelator
-    T(zero(correlator.accumulators))
+function zero(correlator::AbstractCorrelator)
+    update_accumulator(correlator, zero(correlator.accumulators))
 end
 
 """
@@ -154,8 +205,8 @@ $(SIGNATURES)
 
 Filter the Correlator by the function `post_corr_filter`
 """
-function apply(post_corr_filter, correlator::T) where T <: AbstractCorrelator
-    (T.name.wrapper)(map(post_corr_filter, get_accumulators(correlator)))
+function apply(post_corr_filter, correlator::AbstractCorrelator)
+    update_accumulator(correlator, map(post_corr_filter, get_accumulators(correlator)))
 end
 
 """
@@ -167,32 +218,19 @@ ordered from latest to earliest replica.
 """
 function get_correlator_sample_shifts(
     system::AbstractGNSS,
-    correlator::AbstractCorrelator{T, <:SVector{M}},
-    sampling_frequency,
-    preferred_code_shift
-) where {T,M}
-    num_corrs = floor(Int, M / 2)
-    sample_shift = preferred_code_shift_to_sample_shift(
-        preferred_code_shift,
-        sampling_frequency,
-        system
-    )
-    SVector{M}(-num_corrs:num_corrs) .* sample_shift
-end
-function get_correlator_sample_shifts(
-    system::AbstractGNSS,
-    correlator::AbstractCorrelator,
+    num_correlators::Int,
     sampling_frequency,
     preferred_code_shift
 )
-    num_corrs = floor(Int, get_num_accumulators(correlator) / 2)
+    half_num_correlators = num_correlators >> 1
     sample_shift = preferred_code_shift_to_sample_shift(
         preferred_code_shift,
         sampling_frequency,
         system
     )
-    (-num_corrs:num_corrs) .* sample_shift
+    (-half_num_correlators:half_num_correlators) .* sample_shift
 end
+
 function preferred_code_shift_to_sample_shift(
     preferred_code_shift,
     sampling_frequency,
@@ -210,21 +248,16 @@ Calculates the index for the early and late samples
 function get_early_late_index_shift(
     system,
     correlator_sample_shifts,
-    correlator,
     sampling_frequency,
     preferred_code_shift
 )
-    preferred_code_phase_distance = Inf
-    early_late_index_shift = 1
-    for i = get_prompt_index(correlator_sample_shifts) + 1:length(correlator_sample_shifts)
-        code_phase_shift = correlator_sample_shifts[i] * get_code_frequency(system) / sampling_frequency
-        preferred_code_phase_distance_temp = abs(code_phase_shift - preferred_code_shift)
-        if preferred_code_phase_distance_temp < preferred_code_phase_distance
-            early_late_index_shift = i - get_prompt_index(correlator_sample_shifts)
-            preferred_code_phase_distance = preferred_code_phase_distance_temp
-        end
-    end
-    early_late_index_shift
+    sample_shift = preferred_code_shift_to_sample_shift(
+        preferred_code_shift,
+        sampling_frequency,
+        system
+    )
+    min_val, min_idx = findmin(abs, correlator_sample_shifts .- sample_shift)
+    min_idx
 end
 
 """
@@ -233,11 +266,10 @@ $(SIGNATURES)
 Calculate the total spacing between early and late correlator in samples.
 """
 function get_early_late_sample_spacing(
-    correlator_sample_shifts,
-    early_late_index_shift
+    correlator::EarlyPromptLateCorrelator
 )
-    correlator_sample_shifts[get_early_index(correlator_sample_shifts, early_late_index_shift)] -
-    correlator_sample_shifts[get_late_index(correlator_sample_shifts, early_late_index_shift)]
+    correlator.shifts[get_early_index(correlator)] -
+    correlator.shifts[get_late_index(correlator)]
 end
 
 """
@@ -254,26 +286,25 @@ $(SIGNATURES)
 Perform a correlation for multi antenna systems
 """
 function correlate(
-    correlator::T,
+    correlator::AbstractCorrelator{1},
     downconverted_signal::AbstractVector,
     code,
-    correlator_sample_shifts,
     start_sample,
     num_samples
-) where {T <: AbstractCorrelator}
+)
     a_re = zero_accumulators(get_accumulators(correlator), downconverted_signal)
     a_im = zero_accumulators(get_accumulators(correlator), downconverted_signal)
     d_re = downconverted_signal.re
     d_im = downconverted_signal.im
     @avx for i = start_sample:num_samples + start_sample - 1
         for j = 1:length(a_re)
-            sample_shift = correlator_sample_shifts[j] - correlator_sample_shifts[1]
+            sample_shift = correlator.shifts[j] - correlator.shifts[1]
             a_re[j] += d_re[i] * code[i + sample_shift]
             a_im[j] += d_im[i] * code[i + sample_shift]
         end
     end
     accumulators_result = complex.(a_re, a_im)
-    T(map(+, get_accumulators(correlator), accumulators_result))
+    update_accumulator(correlator, map(+, get_accumulators(correlator), accumulators_result))
 end
 
 function zero_accumulators(accumulators::SVector, signal)
@@ -289,13 +320,12 @@ $(SIGNATURES)
 Perform a correlation for multi antenna systems
 """
 function correlate(
-    correlator::AbstractCorrelator{<: SVector{N}},
+    correlator::AbstractCorrelator{M},
     downconverted_signal::AbstractMatrix,
     code,
-    correlator_sample_shifts,
     start_sample,
     num_samples,
-) where {N}
+) where {M}
     a_re = zero_accumulators(get_accumulators(correlator), downconverted_signal)
     a_im = zero_accumulators(get_accumulators(correlator), downconverted_signal)
     d_re = downconverted_signal.re
@@ -303,14 +333,20 @@ function correlate(
     @avx for i = start_sample:num_samples + start_sample - 1
         for k = 1:size(a_re, 2)
             for j = 1:size(a_re, 1)
-                shift = correlator_sample_shifts[k] - correlator_sample_shifts[1]
+                shift = correlator.shifts[k] - correlator.shifts[1]
                 a_re[j,k] += d_re[i,j] * code[shift + i]
                 a_im[j,k] += d_im[i,j] * code[shift + i]
             end
         end
     end
+    update_accumulator(correlator, add_to_previous(get_accumulators(correlator), a_re, a_im))
+end
 
-    typeof(correlator)(map(+, get_accumulators(correlator), eachcol(complex.(a_re, a_im))))
+function add_to_previous(accumulators::SVector{NC, <:SVector}, a_re, a_im) where NC
+    SVector{NC, eltype(accumulators)}(map(+, accumulators, eachcol(complex.(a_re, a_im))))
+end
+function add_to_previous(accumulators::Vector{<:SVector}, a_re, a_im)
+    map(+, accumulators, eachcol(complex.(a_re, a_im)))
 end
 
 function zero_accumulators(accumulators::SVector{NC, <:SVector{NA}}, signal) where {NC, NA}