Okay this is quite a big change :D
This restructure tackles multiple issues:

1. Seperation of concerns: Have a look at the new `track` function. There are now only two main function:
 - `downconvert_and_correlate` to downconvert and correlate the signal. Tracking.jl provides such a function, which does the downconversion and correlation on the CPU. However, this could easily be replaced by another implementation that for e.g. works on the GPU
 - `estimate_doppler_and_filter_prompt` to estimate the Doppler for carrier and code and filter the prompt. Tracking.jl provides the conventional PLLs and DLLs functions. Again this could easily be replaced by another implementation.
 2. Support of tracking multiple satellites in one go. With the previous implementation, you have to call `track` for each satellite. Now you can track multiple satellites with a single `track` call. This becomes important if you'd like to implement a Vector Tracking Loop (VTL). This implementation also allows to track satellites from different systems in one go.
 3. Now that you can downconvert and correlate multiple satellites in one go, there is the possibility to implement a more performant kernel for the GPU.
 
I've done some benchmarks:
New implementation:
```julia
# 4000 samples, 4MHz sampling frequency, single satellite, 1 antenna
julia> @btime track(signal, track_state, sampling_frequency)
  8.940 μs (43 allocations: 4.36 KiB)
``` 
Old implementation:
```julia
# 4000 samples, 4MHz sampling frequency, single satellite, 1 antenna
julia> @btime track(signal, state, sampling_frequency)
  8.224 μs (5 allocations: 1.00 KiB)
```
3 Antennas
New implementation:
```julia
# 4000 samples, 4MHz sampling frequency, single satellite, 3 antennas
julia> @btime track(signal_mat, track_state, sampling_frequency)
  12.641 μs (54 allocations: 7.66 KiB)
``` 
Old implementation:
```julia
# 4000 samples, 4MHz sampling frequency, single satellite, 3 antenna
julia> @btime track(signal_mat, state, sampling_frequency)
  16.064 μs (8 allocations: 94.41 KiB)
```

For a single antenna the new implementation is almost on par with the old implementation. For 3 antennas the new implementation is faster. Note that the new implementation provides the possibility to increase the performance even further with a GPU or by using multiple cores of the CPU (currently it uses only a single core, but the implementation could easily be extended to downconvert and correlate each satellite on a different core).

There are still some action items to be done:
- [ ] Add a small comment to each function
- [ ] Update the docs
- [ ] Provide an easy access to the states of each satellite